+++ CHANGELOG.md
@@ -24,6 +24,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
   particles it contains.  Identical in function to `size`.
 - `get_envelope` now also calculates the normalised emittances `emit_xn` and `emit_yn`,
   not just the geometric ones, as was previously the case.
+- Added `overwrite_progress` kwarg to `cpbd.track.track` function which allows
+  tracking progress to optionally be written on new lines rather than
+  overwritten repeatedly using carriage returns. This is particularly useful
+  when logging the output of a running ocelot simulation to avoid extremely long
+  lines and unreadable log files.
+- Inactive processes in `Navigator` instances are now stored in the
+  `inactive_proccses` attribute. This way processes attached to a `Navigator`
+  remain accessible in that for the lifetime of the Navigator.
+- New `ParameterScanner` class in `cpbd.track` for scanning arbitrary parameters
+  in parallel (either using multiprocessing mpi4py). For example one might scan
+  different compression schemes. Results are compiled into a single hdf5 file.
 
 ### Fixed